+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- **History UI**: Fix calendar hour markers and grid lines not correctly recalculating positions when zooming out.
+
 ## [1.11.0] - 2026-09-16
 
 ### Added